QA: Minor compile warning fix

+++ src/Gulden/rpcgulden.cpp
@@ -251,7 +251,7 @@ static UniValue getwitnessinfo(const JSONRPCRequest& request)
         pTipIndex = chainActive.Tip();
     }
 
-    if (!pTipIndex || pTipIndex->nHeight < Params().GetConsensus().pow2Phase2FirstBlockHeight)
+    if (!pTipIndex || (uint64_t)pTipIndex->nHeight < Params().GetConsensus().pow2Phase2FirstBlockHeight)
         return NullUniValue;
 
     if (request.params.size() >= 2)
